giftcards/requirements.txt
anticaptchaofficial==1.0.44
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.24.82
botocore==1.27.82
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
exceptiongroup==1.0.0rc9
fake-useragent==0.1.11
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
h11==0.14.0
helium==3.0.8
html2image==2.0.1
html5lib==1.1
httplib2==0.20.4
idna==3.4
jmespath==1.0.1
lxml==4.9.1
numpy==1.21.6
oauthlib==3.2.1
outcome==1.2.0
pandas==1.3.5
protobuf==4.21.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.3
webencodings==0.5.1
wsproto==1.2.0
